requirements.txt
albumentations==1.3.1
matplotlib==3.8.0
onnx==1.14.1
onnxruntime
onnxruntime-gpu
opencv-python==4.8.1.78
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
